Fix test failures, sync ORM models with Cloud SQL, and align CI with local standards
CRITICAL FIXES (14 test failures resolved):
- Fixed RSS metadata UPDATE query missing WHERE clause (4 tests, production bug)
- Fixed scheduling cadence float conversion in 12-hour window logic (2 tests)
- Fixed gazetteer OR filtering for entity extraction (1 test)
- Fixed mock connection context manager for alembic tests (1 test)
- Fixed telemetry discovery outcome expectations (1 test)
- Fixed content cleaning telemetry mock assertion (1 test)
- Fixed telemetry deployment readiness tests (4 tests)

ORM MODEL SYNCHRONIZATION:
- Rewrote ExtractionTelemetryV2 with 30 correct columns matching Cloud SQL
- Added ContentTypeDetectionTelemetry model (14 columns)
- Added ContentCleaningSession model (13 columns)
- Added HttpErrorSummary model with UNIQUE constraint (8 columns)
- Fixed http_error_summary INSERT to include first_seen column
- Updated ORM tests to match new schema (article_id as string, new fields)
- Unified test fixture for Cloud SQL schema consistency

CODE QUALITY & CI IMPROVEMENTS:
- Ran ruff linting, fixed 2 truly unused imports
- Restored datetime import needed for docstring examples (noqa: F401)
- Added mypy type checking to CI lint job (advisory, continue-on-error: true)
- Created CI_LOCAL_STANDARDS_COMPARISON.md documenting alignment
- Verified CI YAML syntax and mypy functionality

RESULTS:
- 1260 tests passing (up from 1246, +14 fixed)
- 0 tests failing
- Coverage: 80.76% (exceeds 80% threshold)
- All telemetry uses Cloud SQL schema consistently
- CI now matches local static analysis standards (ruff, black, isort, mypy)
- Ready for Cloud SQL deployment

+++ src/crawler/scheduling.py
@@ -114,13 +114,13 @@ def should_schedule_discovery(
     """
     now = now or datetime.utcnow()
 
-    cadence_days: int = 7
+    cadence_days: float = 7.0
     try:
         if source_meta and isinstance(source_meta, dict):
             freq = source_meta.get("frequency") or source_meta.get("freq")
-            cadence_days = int(parse_frequency_to_days(freq))
+            cadence_days = parse_frequency_to_days(freq)
     except Exception:
-        cadence_days = 7
+        cadence_days = 7.0
 
     dbm = db
     last = _get_last_processed_date(dbm, source_id)
